extract SourceLoader
